cambiados datos a float

+++ src/collector/parser.h
@@ -11,14 +11,14 @@
 
 struct host_info {
     char     ip[32];
-    double   cpu_usage;
-    double   cpu_user;
-    double   cpu_system;
-    double   cpu_idle;
-    uint64_t mem_used_mb;
-    uint64_t mem_free_mb;
-    uint64_t swap_total_mb;
-    uint64_t swap_free_mb;
+    float   cpu_usage;
+    float   cpu_user;
+    float   cpu_system;
+    float   cpu_idle;
+    float    mem_used_mb;
+    float mem_free_mb;
+    float swap_total_mb;
+    float swap_free_mb;
     uint64_t last_mem_ms;   // 0 = never received
     uint64_t last_cpu_ms;   // 0 = never received
     bool     active;
